requirements.txt
feedparser==6.0.11
requests==2.31.0
pyyaml==6.0.1
jinja2==3.1.3
python-dateutil==2.9.0
rapidfuzz==3.6.1
requests-oauthlib==1.3.1
flask==3.0.0
flask-cors==4.0.0
